Update: get genre, prediction, and jurnal by user id

+++ src/services/genre.service.js
@@ -1,5 +1,6 @@
 import { v1 } from "uuid";
 import { query } from "../utils/db.utils.js";
+import { getOne as getUser } from "./user.service.js";
 
 const getAll = async () => {
   const data = await query(`SELECT * FROM genre`);
@@ -9,6 +10,16 @@ const getAll = async () => {
 const getOne = async (id) => {
   const data = await query(`SELECT * FROM genre WHERE id = ?`, [id]);
   return data[0];
+};
+
+const getByUserId = async (id) => {
+  if (await getUser(id)) {
+    const data = await query(
+      `SELECT user_genre.id, nama_genre, jenis_genre, nama as nama_user, user_profile_photo FROM user_genre INNER JOIN genre ON genre.id = genre_id INNER JOIN user ON user.id = user_id WHERE user_id = ?`,
+      [id]
+    );
+    return data;
+  } else throw Error(`User not found`);
 };
 
 const createOne = async (params) => {
@@ -70,4 +81,4 @@ const deleteOne = async (id) => {
   return { message, status };
 };
 
-export { getAll, getOne, createOne, updateOne, deleteOne };
+export { getAll, getOne, createOne, updateOne, deleteOne, getByUserId };
